requirements.txt
fastapi
fastapi-mail
uvicorn
pydantic
redis
motor
python-decouple
PyJWT
passlib